requirements.txt
# BadgePatternTool 核心依赖包
# 只包含运行时必需的依赖，减小打包体积

# GUI框架
PySide6>=6.5.0,<6.8.0

# 图像处理
Pillow>=10.0.0,<11.0.0

# PDF生成
reportlab>=4.0.0,<5.0.0

# 开发和构建工具（可选）
# 注意：以下依赖仅在开发时需要，生产环境可忽略
pyinstaller>=5.0.0; extra == 'dev'
